+++ CHANGELOG.md
@@ -14,6 +14,7 @@ interatomic distances;
 procedure;
 - Function `trminitialpoint` to get an initial point for the specified
 transformation model and planar and torsion angle indices;
+- Functions `pdbstrip` and `pdbenrich` to process PDB files;
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
 - Option to consider planar angles added to `trmdistantangleindices`;
 - `trmcreate` creates an alpha carbon-based transformation model if the